Update Handlebars helpers to use the latest from kss-node
The current Handlebars helpers file is outdated. It does not contain the helpers necessary to render a custom template started from `kss-node --init`, hence https://github.com/PhilJ/gulp-kss/issues/11.

It looks to me like this file was just a copy-paste job from a previous version of kss-node; so this pull request is simply another copy-paste job from the current version.

Using this new helpers file worked for me with the default template version that arrives with gulp-kss and also worked with a new custom template generated from `kss-node --init`.